+++ src/KsWare.Presentation.Lite.TestApp/InitializeComponentTest/ContentControl1.xaml.cs
@@ -1,0 +1,23 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Navigation;
+using System.Windows.Shapes;
+
+namespace ns.InitializeComponentTest {
+	/// <summary>
+	/// Interaction logic for ContentControl.xaml
+	/// </summary>
+	public partial class ContentControl1 : ContentControl {
+		public ContentControl1() {
+			InitializeComponent();
+		}
+	}
+}