+++ src/KsWare.Presentation.Lite.TestApp/InitializeComponentTest/Dictionary1.xaml.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Windows;
+
+namespace ns.InitializeComponentTest {
+	public partial class Dictionary1 : ResourceDictionary {
+
+		public Dictionary1() {
+			InitializeComponent();
+		}
+
+	}
+}