+++ src/KsWare.Presentation.Lite/(ViewLocator)/ViewLocatorStrategyBase.cs
@@ -3,14 +3,15 @@
 // Author           : SchreinerK
 // Created          : 2020-01-27
 //
-// Last Modified By : SchreinerK
-// Last Modified On : 2020-01-27
 // ***********************************************************************
 // <copyright file="ViewLocatorStrategyBase.cs" company="KsWare">
 //     Copyright © by KsWare. All rights reserved.
 // </copyright>
 // <summary></summary>
 // ***********************************************************************
+
+using System;
+
 namespace KsWare.Presentation.Lite {
 
 	internal abstract class ViewLocatorStrategyBase : IViewLocatorStrategy {
@@ -21,6 +22,8 @@ namespace KsWare.Presentation.Lite {
 		public string[] ViewModelFolderNames { get; set; } = { "", "ViewModel" };
 
 		public abstract object GetView(object viewModelOrType);
+		public abstract Type GetViewType(object viewModelOrType);
+
 	}
 
 }