+++ ChangeLog.md
@@ -1,10 +1,15 @@
 ﻿# ChangeLog
 KsWare.Presentation.Lite
 
+#### [2020-02-01]
+- UPD ViewLocator
+    1. NEW InitializeComponentWrapper to support views with no code behind for DataTemplete/ControlTemplate
+    2. UPD CreateDataTemplete/CreateControlTemplate
+
 #### [2020-31-01]
 - UPD ViewModelViewConverter supports DataTemplate and ControlTemplateand as TargetType and accepts also a Type as value
 - NEW ViewModelPresenterExtension
-- 
+
 #### [2020-01-30]
 - NEW Binding/BindingBase markup extension
 - NEW DebugBindung markup extension